+++ R/familias-deprecated.R
@@ -1,9 +1,8 @@
 #' Import/export from Familias
 #'
 #' @description
-#' Functions for reading and writing .fam files associated with the Familias
-#' software for forensic kinship computations, and for converting objects from
-#' the R package `Familias` into the `pedsuite`.
+#' Functions for reading .fam files associated with the Familias
+#' software for forensic kinship computations.
 #'
 #' ***Deprecated*** These functions have been moved to a separate package,
 #' `pedFamilias`, and will be removed from `forrel` in a future version.
@@ -21,35 +20,3 @@ readFam = function(...) {
   pedFamilias::readFam(...)
 }
 
-
-#' @rdname familias
-#' @export
-writeFam = function(...) {
-  cat("Deprecated, use `pedFamilias::writeFam()` instead.\n")
-  pedFamilias::writeFam(...)
-}
-
-
-#' @rdname familias
-#' @export
-openFamilias = function(...) {
-  cat("Deprecated, use `pedFamilias::openFamilias()` instead.\n")
-  pedFamilias::openFamilias(...)
-}
-
-
-#' @rdname familias
-#' @export
-Familias2ped = function(...) {
-  cat("Deprecated, use `pedFamilias::Familias2ped()` instead.\n")
-  pedFamilias::Familias2ped(...)
-}
-
-
-
-#' @rdname familias
-#' @export
-readFamiliasLoci = function(...) {
-  cat("Deprecated, use `pedFamilias::readFamiliasLoci()` instead.\n")
-  pedFamilias::readFamiliasLoci(...)
-}