+++ R/forrel-package.R
@@ -4,5 +4,6 @@
 ## usethis namespace: start
 #' @import pedtools
 #' @import pedprobr
+#' @importFrom pbapply pblapply pboptions
 ## usethis namespace: end
 NULL